use parseFloat instead of parseInt for numbers

+++ src/client/app/shared/json-data-view-model/data-view-translator.ts
@@ -18,7 +18,7 @@ export class DataViewTranslatorService {
       let viewTypedPointer: TypedPointer = fieldMap.viewTypedPointer;
       let dataValue: string = JsonPointerService.get(dataModel, dataTypedPointer.pointer);
       if (viewTypedPointer.type === 'number') {
-        JsonPointerService.set(viewModel, viewTypedPointer.pointer, parseInt(dataValue));
+        JsonPointerService.set(viewModel, viewTypedPointer.pointer, parseFloat(dataValue));
       } else {
         JsonPointerService.set(viewModel, viewTypedPointer.pointer, dataValue);
       }
@@ -38,7 +38,7 @@ export class DataViewTranslatorService {
       let viewTypedPointer: TypedPointer = fieldMap.viewTypedPointer;
       let viewValue: string = JsonPointerService.get(viewModel, viewTypedPointer.pointer);
       if (dataTypedPointer.type === 'number') {
-        JsonPointerService.set(dataModel, dataTypedPointer.pointer, parseInt(viewValue));
+        JsonPointerService.set(dataModel, dataTypedPointer.pointer, parseFloat(viewValue));
       } else {
         JsonPointerService.set(dataModel, dataTypedPointer.pointer, viewValue);
       }
